Initial commit
* Single screen contains:
	- Toggleable switch for enabling/disabling device Bluetooth
	- Clickable button for start/stop scan for bluetooth devices.
	- Gridview linked to a custom adapter prepared to show each discovered
device.
* The app tests and asks for permission to enable bluetooth(redundant with the
switch but I'm not sure if the switch will be permanent).
* StartScan and StopScan functions.
* Custom code filters the discovered devices for the ones who advertise the
expected service.
* The adapter stores arrayLists of ScanResults, Boolean(Connected, DoorState)
and BluetoothGattCharacteristic(DoorContact, DoorStrike, DoorReq).
* The BluetoothLeService regulates all interactions with the devices, right
now that means that when a connection is called, the service will create a new
BluetoothGatt to connect to the device and keep track of which one is which
for purposes of interactions using the expected callbacks that come from the
BluetoothGatt library.
* (De-activated) Single connection using a single gatt saved in the service.
* (De-activated) StartActivity View used for Bluetooth permission.
* (De-activated) SearchActivity View used for device discovery.
* (De-activated) DetailsActivity View used for connection and interaction with
a device selected in the SearchActivity.

* (Currently in progress) The transition from single Gatt/Devices to multiple
is still ongoing, the broadcast information used for communication with the
main activity has been changed, that is why some previous functions of pre-
vious activities do not work anymore.
* (Currently in progress) Keeping up with multiple devices in the gridView.

* (To be Done) Better layout of the gridView for number of devices growth is
needed, right now only the height dinamic growth has worked, might need to use
a custom function for that.